// src/FE/src/api/generated/models/jobLinkInfoResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Workslip.Api | v1
 * OpenAPI spec version: 1.0.0
 */

export interface JobLinkInfoResponse {
  id: string;
  linkedReportId: string;
  linkedReportNumber: string;
  linkedCustomerName: string;
  linkedAddress: string;
  linkedStatus: string;
}