// src/FE/src/api/generated/models/jobReportSummaryViewModel.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Workslip.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { AssignedUserResponse } from './assignedUserResponse';
import type { CustomerSnapshotResponse } from './customerSnapshotResponse';
import type { JobLinkInfoResponse } from './jobLinkInfoResponse';
import type { JobReportSummaryObservationResponse } from './jobReportSummaryObservationResponse';
import type { JobReportSummaryWorkResponse } from './jobReportSummaryWorkResponse';
import type { JobStatus } from './jobStatus';
import type { WorksheetResponse } from './worksheetResponse';

export interface JobReportSummaryViewModel {
  id: string;
  organizationId: string;
  /** @nullable */
  reportNumber: string | null;
  status: JobStatus;
  /** @nullable */
  customerId: string | null;
  customerSnapshot: CustomerSnapshotResponse;
  /** @nullable */
  destinationAddress: string | null;
  /** @nullable */
  destinationZipCode: string | null;
  /** @nullable */
  destinationCity: string | null;
  work: JobReportSummaryWorkResponse;
  observations: JobReportSummaryObservationResponse;
  links: JobLinkInfoResponse[];
  assignedUsers: AssignedUserResponse[];
  worksheets: WorksheetResponse[];
  /**
     * @nullable
     * @pattern ^-?(?:0|[1-9]\d*)(?:\.\d+)?$
     */
  totalHours: number | string | null;
  /**
     * @nullable
     * @pattern ^-?(?:0|[1-9]\d*)$
     */
  totalOutlay: number | string | null;
  softDeleted: boolean;
  jobType: string;
  /** @nullable */
  rejectionNote: string | null;
  /** @nullable */
  createdJobIds?: string[] | null;
}